requirements.txt
certifi==2023.7.22
datetime==5.1
Flask==2.0.3
Flask-API==3.1
gunicorn==21.2.0
machaao==0.3.5
PyJWT==2.3.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.24.0
typing_extensions==4.8.0
language_tool_python==2.7.1
urllib3==1.25.11
Werkzeug==2.2.2
replicate==0.25.2
snips-nlu==0.20.2
